Remove outdated exponent functions and messages
